requirements.txt
streamlit
streamlit_chat
langchain
sentence_transformers
openai
unstructured
unstructured[local-inference]
pinecone-client
python-dotenv
pypdf
chromadb
tiktoken